Add license validation call when disabling app
This is needed for HA environments. In HA setup, the instance
that first validates the license for an app will propagate its
status change to INVALID_LICENSE_DISABLED for the other instances
in the cluster, but other instances would not have the error messages
detailing why the license is invalid.

So we added a license validation call to this moment for the instance
to be able to report the errors if needed