Add missing unregister component method (#144)
* chore: change comment style to obtain VS Code Intellisense

* feat: add unregisterComponent method

We made follwing changes for the AppExternalComponentManager:
* add new private property - appTouchedComponents
* add new public method unregisterComponent
* add new public method getExternalComponent

* Add unregisterComponent in the lifecycle methods

We need to call unregisterComponent method to remove the corresponding external  components from the providedComponents when following AppManager's methods are called:
*  `unload` is called
* `disable` is called <-> Disable an app
*  `remove` is called <-> Uninstall an app
*  `updateAppsMarketplaceInfo` is called and validateLicense not pass
* `initializeApp` is called and validateLicense not pass
* `enableApp` is called but is not success

* Add new AppExternalComponentManager method purgeComponent

when remove an app from the system, we need delete its external component from both providedComponents map set and appTouchedCompoents map set.

* fix: new deployed app registers externalComponents automatically

since the new deployed app will try to register their externalComponent when the  `initializeApp` method of AppManager is called. This is not expected behavior. We introduce a new parameter `addToProvidedComponents` for the method `registerComponent` to handle the above problem. Without specifying `registerComponent` to true, the app only "declares" it touched the external component when it is initialized. Only if the `enableApp` method is called, the external component will truly added to providedComponents map set, which means we can access it in the `Game Center`.